+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1](https://github.com/LebJe/TOMLKit/releases/tag/0.1.0) - 2021-06-17
+
+### Added
+
+-   Add `TOMLArrayIterator` and `TOMLTableIterator`.
+
+### Fixed
+
+-   `TOMLArray.checkIndex(index:)` now checks that `index <= self.endIndex - 1` instead of `index <= self.endIndex`.
+
 ## [0.1.0](https://github.com/LebJe/TOMLKit/releases/tag/0.1.0) - 2021-06-17
 
 ### Added